Refactoring based on code review comments and update file directories
