+++ src/pages/experiments/[experimentId]/data-processing/index.jsx
@@ -22,7 +22,7 @@ import MitochondrialContent from './filter-cells/components/MitochondrialContent
 import Classifier from './filter-cells/components/Classifier/Classifier';
 import GenesVsUMIs from './filter-cells/components/GenesVsUMIs/GenesVsUMIs';
 import DoubletScores from './filter-cells/components/DoubletScores/DoubletScores';
-import DataIntegration from './data-integration/components/DataIntegration/DataIntegration';
+import DataIntegration from './data-integration/components/DataIntegration';
 import EmbeddingPreview from './configure-embedding/components/EmbeddingPreview';
 
 const { Text } = Typography;
